more changes to id and functions in Activity Package

+++ Sem4_Sprint1Two_Main/src/main/java/com/keyin/activity/ActivityService.java
@@ -187,7 +187,7 @@ public class ActivityService {
         Map<String, Object> parameters = activity.getParameters();
 
         Aircraft newAircraft = new Aircraft();
-        newAircraft.setId((Integer) parameters.get("id"));
+        newAircraft.setId((Long) parameters.get("id"));
         newAircraft.setType((String) parameters.get("type"));
         newAircraft.setAirlineName((String) parameters.get("airlineName"));
         newAircraft.setNoOfPassengers((Integer) parameters.get("numberOfPassengers"));
@@ -201,7 +201,7 @@ public class ActivityService {
         Map<String, Object> parameters = redoActivity.getParameters();
 
         Aircraft newAircraft = new Aircraft();
-        newAircraft.setId((Integer) parameters.get("id"));
+        newAircraft.setId((Long) parameters.get("id"));
         newAircraft.setType((String) parameters.get("type"));
         newAircraft.setAirlineName((String) parameters.get("airlineName"));
         newAircraft.setNoOfPassengers((Integer) parameters.get("numberOfPassengers"));
@@ -215,7 +215,7 @@ public class ActivityService {
         Map<String, Object> parameters = activity.getParameters();
 
         Airport newAirport = new Airport();
-        newAirport.setId((Integer) parameters.get("id"));
+        newAirport.setId((Long) parameters.get("id"));
         newAirport.setName((String) parameters.get("name"));
         newAirport.setCityId((Integer) parameters.get("cityId"));
 
@@ -227,7 +227,7 @@ public class ActivityService {
         Map<String, Object> parameters = redoActivity.getParameters();
 
         Airport newAirport = new Airport();
-        newAirport.setId((Integer) parameters.get("id"));
+        newAirport.setId((Long) parameters.get("id"));
         newAirport.setName((String) parameters.get("name"));
         newAirport.setCityId((Integer) parameters.get("cityId"));
 
@@ -239,7 +239,7 @@ public class ActivityService {
         Map<String, Object> parameters = activity.getParameters();
 
         City newCity = new City();
-        newCity.setId((Integer) parameters.get("id"));
+        newCity.setId((Long) parameters.get("id"));
         newCity.setName((String) parameters.get("name"));
         newCity.setProvince((String) parameters.get("province"));
         newCity.setPopulation((Integer) parameters.get("population"));
@@ -252,7 +252,7 @@ public class ActivityService {
         Map<String, Object> parameters = redoActivity.getParameters();
 
         City newCity = new City();
-        newCity.setId((Integer) parameters.get("id"));
+        newCity.setId((Long) parameters.get("id"));
         newCity.setName((String) parameters.get("name"));
         newCity.setProvince((String) parameters.get("province"));
         newCity.setPopulation((Integer) parameters.get("population"));
@@ -265,7 +265,7 @@ public class ActivityService {
         Map<String, Object> parameters = activity.getParameters();
 
         Passenger newPassenger = new Passenger();
-        newPassenger.setId((Integer) parameters.get("id"));
+        newPassenger.setId((Long) parameters.get("id"));
         newPassenger.setFirstname((String) parameters.get("firstname"));
         newPassenger.setLastName((String) parameters.get("lastName"));
         newPassenger.setPhoneNumber((String) parameters.get("phoneNumber"));
@@ -280,7 +280,7 @@ public class ActivityService {
         Map<String, Object> parameters = redoActivity.getParameters();
 
         Passenger newPassenger = new Passenger();
-        newPassenger.setId((Integer) parameters.get("id"));
+        newPassenger.setId((Long) parameters.get("id"));
         newPassenger.setFirstname((String) parameters.get("firstname"));
         newPassenger.setLastName((String) parameters.get("lastName"));
         newPassenger.setPhoneNumber((String) parameters.get("phoneNumber"));
